+++ app/src/main/java/dev/upaya/shf/ui/session/SessionNavigation.kt
@@ -1,11 +1,11 @@
-package dev.upaya.shf.ui.exercises
+package dev.upaya.shf.ui.session
 
 import androidx.navigation.NavController
 import dev.upaya.shf.exercises.exerciselist.ExerciseID
 import dev.upaya.shf.exercises.exerciselist.ExerciseRoute
 
 
-fun NavController.navigateToExercise(
+fun NavController.navigateToExerciseSession(
     exerciseRoute: ExerciseRoute,
     exerciseID: ExerciseID,
 ) {
